Rename integration client tests to underscore naming
