+++ integration_client/__init__.py
@@ -1,1 +1,1 @@
-"""Test utilities package for integration-client"""
+"""Test utilities package for integration_client"""
